+++ Demo/src/main/java/com/braintreepayments/demo/Settings.java
@@ -202,9 +202,6 @@ public class Settings {
         return getPreferences(context).getBoolean("vault_venmo", true);
     }
 
-    public static boolean venmoFallbackToWeb(Context context) {
-        return getPreferences(context).getBoolean("venmo_fallback_to_web", false);
-    }
     public static boolean isAmexRewardsBalanceEnabled(Context context) {
         return getPreferences(context).getBoolean("amex_rewards_balance", false);
     }