+++ Venmo/src/test/java/com/braintreepayments/api/VenmoClientUnitTest.java
@@ -15,6 +15,9 @@ import android.net.Uri;
 
 import androidx.test.core.app.ApplicationProvider;
 
+import junit.framework.Assert;
+import junit.framework.TestCase;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.Before;