+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Venmo
   * Add `setIsFinalAmount()` to `VenmoRequest`
 * ThreeDSecure
-  * Call cleanup method to resolve `Cardinal.getInstance` memory leak (fixes #898)
+  * Call cleanup method to resolve `Cardinal.getInstance` memory leak
 
 ## 4.41.0 (2024-01-18)